requirements.txt
opencv-python==4.8.1.78
pyautogui==0.9.54
numpy==1.24.3
mss==9.0.1
Pillow==10.0.1
keyboard==0.13.5
pynput==1.7.6
colorama==0.4.6 